adds ResendLEDState message to core and calls it from MantaFlext
This makes it so that every time the manta is connected it sends the
accumulated buffer of LEDs and control bits. This means both that if
you unplug and re-plug the manta in it will maintain the same state, and
also if you do some init in your patch from a loadbang before the manta
is connected it should still catch them.

closes #12