adds ability to force queuing of an outgoing frame
Normally, calling WriteFrame several times without waiting for the
messages to go down will cause the messages to get merged together.
sometimes however, we want to make sure that an outgoing message
doesn't get merged with any messages that are currently in the queue.
For instance, we want to make sure that a SetLEDControl message gets
sent AFTER any pending LED messages. This commit adds a "forceQueued"
parameter to WriteFrame to control this behavior.